# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anophel API Documentation",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get All Doctors",
                "description": "Retrieve all doctors",
                "operationId": "f2faa955668b3e6887792008f2d09661",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "doctor_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "specialization_id_fk": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "doctor_rooms": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "doctor_room_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "room": {
                                                            "type": "string",
                                                            "example": "151"
                                                        },
                                                        "doctor_id_fk": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get Doctor by ID",
                "description": "Retrieve a doctor by its ID",
                "operationId": "800e90f10564a8eec88cd8284dac12a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "doctor": {
                                            "properties": {
                                                "doctor_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "specialization_id_fk": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "doctor_rooms": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "doctor_room_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "room": {
                                                                "type": "string",
                                                                "example": "151"
                                                            },
                                                            "doctor_id_fk": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete Doctor by ID",
                "description": "Delete a doctor by its ID",
                "operationId": "0ce07cef46a97c7cf00c188f3d236076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "string",
                                            "example": "دکتر مورد نظر با موفقیت حذف شد."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
}